Added Checkout and Grand Total
And some visual changes

+++ VendorService/VendorService/Models/Vendor.cs
@@ -11,8 +11,8 @@ namespace VendorService.Models
         [Key]
         public int Id { get; set; }
         public string Name { get; set; }
-        public double DeliveryCharge { get; set; }
-        public double Rating { get; set; }
+        public int DeliveryCharge { get; set; }
+        public int Rating { get; set; }
 
     }
 }
